Add a link to C# port of Shield style badge lib.
When we try to add a badge service to [visual studio gallery](visualstudiogallery.msdn.microsoft.com), we find Shield is the best option but there is no .NET implementation. So we port it to [.NET](https://github.com/rebornix/DotBadge) for better code integration.

I'll appreciate it if you can add a link to C# port of Shield badge library and CLI, which can bring more audience to this tiny draft project. But we can understand if you don't, as it's not an official one :)